backend/requirements.txt
<<<<<<< HEAD

=======
>>>>>>> b46545bf3 (Add build and dist folders)
setuptools>=65.5.0
wheel

#Backend
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.2
plotly==5.17.0
pydantic==2.5.0
apscheduler==3.10.4
websockets==12.0
python-dateutil==2.8.2
alembic==1.12.1
paho-mqtt==1.6.1
aiohttp==3.9.1
aiocoap==0.4.3
asyncio-mqtt==0.13.0
dnspython==2.6.1
email-validator==2.3.0
streamlit-custom-checkbox-rpathak==0.1.0
python-http-client==3.3.7
sendgrid==6.12.5
werkzeug==3.1.3
<<<<<<< HEAD
=======
psutil==5.9.8
pywebview==4.1
requests==2.31.0
>>>>>>> b46545bf3 (Add build and dist folders)


# Frontend
streamlit==1.50.0
requests==2.31.0
pandas==2.0.3
plotly==5.17.0
<<<<<<< HEAD
python-dotenv==1.0.0
=======
python-dotenv==1.0.0
>>>>>>> b46545bf3 (Add build and dist folders)
